Fix building from sdist/mapnik-config + include headers

+++ setup.py
@@ -40,8 +40,18 @@ if os.environ.get('SYSTEM_MAPNIK'):
      f_paths.write("inputpluginspath = '{path}'\n".format(path=input_plugin_path))
      f_paths.write("fontscollectionpath = '{path}'\n".format(path=font_path))
 else:
-     os.symlink(bin_path, 'packaging/mapnik/bin')
-     os.symlink(lib_path, 'packaging/mapnik/lib')
+     if not os.path.exists('packaging/mapnik/bin'):
+          os.symlink(bin_path, 'packaging/mapnik/bin')
+     if not os.path.exists('packaging/mapnik/lib') :
+          os.symlink(lib_path, 'packaging/mapnik/lib')
+     else:
+          names = (name for name in os.listdir(lib_path) if os.path.isfile(os.path.join(lib_path, name)))
+          for name in names:
+               if not os.path.exists(os.path.join('packaging/mapnik/lib', name)):
+                    os.symlink(os.path.join(lib_path, name), os.path.join('packaging/mapnik/lib', name))
+          input_plugin_path = check_output([mapnik_config, '--input-plugins'])
+          if not os.path.exists('packaging/mapnik/lib/mapnik/input'):
+               os.symlink(input_plugin_path, 'packaging/mapnik/lib/mapnik/input')
      f_paths.write("mapniklibpath = os.path.join(os.path.dirname(__file__), 'lib')\n")
      f_paths.write("inputpluginspath = os.path.join(os.path.dirname(__file__), 'lib/mapnik/input')\n")
      f_paths.write("fontscollectionpath = os.path.join(os.path.dirname(__file__), 'lib/mapnik/fonts')\n")
@@ -124,9 +134,11 @@ if os.environ.get("CXX", False) == False:
 
 setup(
      name="mapnik",
+     include_package_data=True,
      packages=find_namespace_packages(where="packaging"),
      package_dir={"": "packaging"},
      package_data={
+          "mapnik.include": ["*.hpp"],
           "mapnik.bin": ["*"],
           "mapnik.lib": ["libmapnik*"],
           "mapnik.lib.mapnik.fonts":["*"],
